add statedata and refactor mapcontainer
